Render new mappings component on the evaluations frontend, add function with logic to migrate legacy
